+++ packages/logger/cypress/integration/logger.test.ts
@@ -1,6 +1,7 @@
-import { isValidateTimestamp } from "./utils";
-
 const { cy, describe, it } = global;
+
+const isValidateTimestamp = text =>
+  /^([0-9]{2})\:([0-9]{2}):([0-9]{2})$/.test(text);
 
 describe("Logger TCs", () => {
   it("Events are logged (check the format)", () => {
@@ -11,7 +12,7 @@ describe("Logger TCs", () => {
         row.find("[data-cy=logger-arrows-green]");
 
         expect(
-          isValidateTimestamp(row.find("[data-cy=logger-timestamp]"))
+          isValidateTimestamp(row.find("[data-cy=logger-timestamp]").text())
         ).to.eq(true);
 
         console.log(row.get("[data-cy=logger-title]"));
