+++ packages/logger/package.json
@@ -9,9 +9,9 @@
   },
   "license": "MIT",
   "scripts": {
-    "build": "parcel build index.ts --out-dir lib",
+    "build": "rimraf lib && cross-env BABEL_ENV=commonjs babel src/index.ts --out-dir lib --copy-files",
     "cypress": "cypress run",
-    "start": "parcel start index.html",
+    "start": "parcel start src/example/index.html",
     "lint": "tslint -p tsconfig.json --fix",
     "prettier": "prettier --write '{src,tests}/**/*.{ts,tsx}'",
     "test": "npm-run-all -p -r start cypress"
@@ -23,9 +23,11 @@
   },
   "devDependencies": {
     "@types/node": "^11.13.4",
+    "cross-env": "^5.2.0",
     "cypress": "^3.2.0",
     "npm-run-all": "^4.1.5",
     "parcel-bundler": "^1.12.3",
+    "rimraf": "^2.6.3",
     "tslint": "^5.15.0",
     "typescript": "^3.4.3"
   }
