+++ packages/logger/src/index.ts
@@ -1,0 +1,3 @@
+export { Logger } from "./logger";
+export { LoggerProps } from "./types";
+export { Event } from "./types";
