+++ packages/logger/src/types.ts
@@ -1,3 +1,5 @@
+import { Observable } from "rxjs";
+
 export interface Event {
   timestamp: number;
   correlationId: string;
@@ -5,4 +7,8 @@ export interface Event {
   serviceName: string;
   methodName: string;
   data: any;
+}
+
+export interface LoggerProps {
+  logs$: Observable<Event>;
 }
